# Implement session pooling, add thread safety fixes and optimizations