--- a/src/libp11-int.h
+++ b/src/libp11-int.h
@@ -48,7 +48,6 @@ typedef struct pkcs11_ctx_private {
 	void *ui_user_data;
 	unsigned int forkid;
 	PKCS11_RWLOCK rwlock;
-	int sign_initialized;
 } PKCS11_CTX_private;
 #define PRIVCTX(ctx)		((PKCS11_CTX_private *) ((ctx)->_private))