--- a/src/libp11-int.h
+++ b/src/libp11-int.h
@@ -49,7 +49,6 @@ typedef struct pkcs11_ctx_private {
 	unsigned int forkid;
 	PKCS11_RWLOCK rwlock;
 	int sign_initialized;
-	int decrypt_initialized;
 } PKCS11_CTX_private;
 #define PRIVCTX(ctx)		((PKCS11_CTX_private *) ((ctx)->_private))