Improve performance and improve safety in multi threaded usafe of OpenSSL.
See individual commit messages for more detailed explanation of each change.

E.g. BIND9 dnssec-signzone test before these changes (on my i7-8 laptop):
```
Signatures generated:                    10216
Signing time in seconds:                13.666
Signatures per second:                 747.522
Runtime in seconds:                     13.726

real	0m 13.73s
user	0m 14.24s
sys	0m 1.25s

```

And after:
```
Signatures generated:                    10216
Signing time in seconds:                 4.779
Signatures per second:                2137.268
Runtime in seconds:                      4.833

real	0m 4.84s
user	0m 21.88s
sys	0m 2.53s
```