@mtrojnar 2021-03-22
The performance of you PKCS#11 token is really impressive.  Which device do you use?
Have you verified that the produced signatures are correct?  Have you tried any other devices?