@fabled 2021-03-22
> The performance of you PKCS#11 token is really impressive. Which device do you use?

Results are from softhsm2 (openssl build) with RSA 2K keys.

> Have you verified that the produced signatures are correct? Have you tried any other devices?

I did thorough analysis of the libp11 and relevant openssl code. I have quite a bit of experience implementing PKCS#11 tokens, and using them. So special attention has been paid to correctness. And as documented, this also fixes several correctness issues. There's still some potential fixes pending, but will address them in future PRs.

This PR passes test suite as correctness proof. And running `dnssec-signzone` with `-a` to verify the signature will confirm all signatures are good:
```
Signatures successfully verified:        10216
```

Other devices are not yet tested, but is planned against a high-end external HSM device. Can report back when results are available.