@dengert 2021-03-22
This PR should also be tested against low tech devices and other PKCS\#11 modules, that may not have thread protection, depending on libp11 single threading calls to these modules.

This may not be a problem, as I don't see any changes to pkcs11_CTX_load where C_Initialize is called and return values are checked. But based on other discussions, the intent of this change it to queue requests for crypto operations for the same key as close to the hardware as possible, which could be in a card driver or the card itself, possible by using multiple channels to the card. 

https://github.com/johughes99  on OpenSC lists you have said you have been testing PKCS\#11 modules. Do you have any comments on the PR?