@fabled 2021-03-22
Yes, further testing is definitely needed.

Regarding the assumptions made, the PKCS#11 v2.40 states the following:
 - Base Spec 5.4 C_Initialize:
    >A call to C_Initialize specifies one of four different ways to support multi-threaded access via the value of the CKF_OS_LOCKING_OK flag in the flags field and the values of the CreateMutex, DestroyMutex, LockMutex, and UnlockMutex function pointer fields:
    >...
    > 2.     If the flag is set, and the function pointer fields aren’t supplied (i.e., they all have the value NULL_PTR), that means that the application will be performing multi-threaded Cryptoki access, and the library needs to use the native operating system primitives to ensure safe multi-threaded access.  If the library is unable to do this, C_Initialize should return with the value CKR_CANT_LOCK.
    And We pass this flag, so the library should return error if it does not support errors.
 - Usage Guide 2.5.3
   > In particular, when an application initializes a Cryptoki library with a call to C_Initialize, it can specify one of four possible multi-threading behaviors for the library:
   > ...
   > 2.       The application can specify that it will be accessing the library concurrently from multiple threads, and the library must be able to use native operation system synchronization primitives to ensure proper thread-safe behavior.
 - Usage Guide 4.1
   > Different threads of an application should never share sessions, unless they are extremely careful not to make function calls at the same time.  This is true even if the Cryptoki library was initialized with locking enabled for thread-safety.

Of course older libraries might comply to older versions of this. But my current understanding is that, also with the older spec versions `CKF_OS_LOCKING_OK` means the library must go into multi-threaded mode or return error.

And as noted in the Usage Guide, the rule is that multi threaded access must never use same session in multiple threads concurrently. Session is the per-thread context concept. Same session can be used in multiple threads though when application synchronizes things so that it's not used concurrently. This is the approach selected in this PR: a session pool is maintained and sessions are checked out when needed, and returned after use.

The Usage Guide is also very clear that same object handles can are shareable between sessions.

This was the simplest way to implement proper multithreaded support. And it conforms to the PKCS#11 standard. If some PKCS#11 module breaks, it is likely non-conforming and should be fixed. We can of course add additional knobs to adjust the session pool size so that a single threaded access is forced globally if needed.

But obviously this PR is a non-trivial, complex and in some parts intrusive change set. So thorough review and testing is needed. I am sure there are points were changes will be required. I tried to put most thing in isolated individual commits so it might make sense to review and merge this in sets. There are also few fairly large commits that refactor some specific API.

And as explained in the commits, this series even fixes some issues where the single session was "accidentally" used concurrently from multiple threads. Causing crashes such as #393.

I have low-end HSM device here, soon two different ones. And will be testing them soon. But wanted to share this early for review. I am happy to fix any issues found.