@dengert 2021-03-22
Thanks, for the quick response. 

The way I read PKCS\#11 v2.20, is if CK_C_INITIALIZE_ARGS_PTR
is NULL, "the application will not be accessing Cryptoki through multiple threads simultaneously." If it is not NULL, then the application is asking if the module can support
either CKF_OS_LOCKING_OK or use the routines provided by the application.  
If the module can't do one of these it should return CKR_CANT_LOCK
(Or return CKR_NEED_TO_CREATE_THREADS if the module needs to use threads and the application told it to not use threads via CKF_LIBRARY_CANT_CREATE_OS_THREADS). 

So that should not be a problem, as current libp11 always passes in CKF_OS_LOCKING_OK, but  I have not looked closely how libp11 actually uses threads. 

Looks like you are addressing this concern.