@mtrojnar 2021-03-23
We are testing it with stunnel. I should have some news soon.