@fabled 2021-03-24
AEP Keyper HSM PKCS#11 (with load-balancer support) was tested now. No bugs or issues found. Performance differences as follows (with similar dnssec-signzone test):
1. RSA keys: before patch ~250 signatures/s, with PR 1333 signatures/s
2. EC keys: before PR ~210 signatures/s, with PR applied 615 signatures/s

Showing 3-5x speedup.

Testing did reveal that the operations are not distributed evenly to the load balancing cluster. So I will write a follow up commit to change the current LIFO strategy to FIFO in the session pooling.

Also comparison against other software implies there is still potential for increased performance. I will do some additional analysis if something can be still improved in the hot crypto paths, and followup with separate commits as needed.