@mtrojnar 2021-04-08
The initial testing results are quite positive.  We could not observe any speedup with stunnel, but I guess the usage pattern in TLS might be different from dnssec-signzone.

The one change I don't like is the replacement of the portable OpenSSL synchronization API with pthreads.  The rationale for using OpenSSL was that no additional changes are needed to libp11 after OpenSSL and the PKCS#11 module are ported to the new platform (possibly, an obscure embedded OS). Also, I don't really like the idea of emulating pthreads API with the Windows API.  Is there a significant performance gain of using pthreads directly?  Would you consider reverting this part of your PR?

Otherwise, I really like the code and I'll be glad to merge your PR.